Refactor into GamePlay component
